+++ bentoml/archive.py
@@ -25,7 +25,7 @@ import uuid
 from datetime import datetime
 
 from bentoml.utils import Path
-from bentoml.utils.module_helper import copy_used_py_modules
+from bentoml.utils.py_module_utils import copy_used_py_modules
 from bentoml.utils.exceptions import BentoMLException
 from bentoml.utils.s3 import is_s3_url, upload_to_s3, download_from_s3
 from bentoml.loader import load_bentoml_config
